Updates favicon and improves UI
Updates the favicon to a todo list icon for better visual representation.

Improves UI by adjusting input width for better responsiveness and adds margin to main container.

Fixes a minor issue where the `saveToLocalStorage` function was unnecessarily being passed as a dependency to the `useEffect` hook in the `Todo` component, preventing unnecessary re-renders.